backend/quantum_api/ml_predictor.py
import tensorflow as tf
import numpy as np

def predict_quantum_state(circuit_data):
    # Example: A simple TensorFlow model for prediction
    model = tf.keras.Sequential([
        tf.keras.layers.Dense(64, activation='relu'),
        tf.keras.layers.Dense(2, activation='softmax')
    ])
    model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])

    # Dummy data for training (replace with real data)
    x_train = np.random.rand(100, len(circuit_data))
    y_train = np.random.randint(2, size=(100, 2))

    # Train the model
    model.fit(x_train, y_train, epochs=10)

    # Predict the output state
    prediction = model.predict(np.array([circuit_data]))
    return prediction
